refactor(mm-next): move component `story/wide/credits` into shared component
